requirements.txt
mkdocs-material==8.1.9
mkdocs-awesome-pages-plugin==2.6.1
importlib-metadata==4.10.1
gitpython==3.1.30